plate-detector-simulator/requirements.txt
dnspython==2.3.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
python-etcd==0.4.5
urllib3==2.0.3
